[#39] move: facade 패키지 생성 및 facade 파일 이동
